libs/common/requirements.txt
pydantic>=2.0.0
asyncpg>=0.28.0
nats-py[nkeys]>=2.5.0
